tests: use vitest for discogs tests
