+++ tests/discogs/README.md
@@ -6,11 +6,12 @@ Snapshot tests for `parseDiscogsRelease` in `discogs_importer.user.js`. Each tes
 
 ```
 tests/discogs/
-  config.ts       release URLs to test
-  load-parser.ts  loads parseDiscogsRelease in Node
-  run.ts          test runner
-  fixtures/       raw Discogs API responses
-  snapshots/      expected parser output
+  config.ts            releases to test (URL + description)
+  discogs.test.ts      Vitest test file
+  fetch-fixtures.ts    fetches fixtures from the Discogs API
+  load-parser.ts       loads parseDiscogsRelease in Node
+  fixtures/            raw Discogs API responses
+  snapshots/           expected parser output
 ```
 
 ## Fixtures and snapshots
@@ -19,32 +20,34 @@ A **fixture** is the JSON returned by the Discogs API for a release (e.g. `fixtu
 
 A **snapshot** is the JSON produced by running `parseDiscogsRelease` on that fixture (e.g. `snapshots/1996829.json`). It is the expected MusicBrainz import data: artist credits, labels, discs, tracks, durations, and so on.
 
+Snapshot files are excluded from oxfmt formatting so their format is controlled exclusively by the test runner.
+
 When you change the parser, tests re-parse the fixtures and compare the result to the snapshots.
 
 ## How tests work
 
-1. Read each URL listed in `config.ts`.
+1. Read each entry listed in `config.ts`.
 2. Load the matching fixture from `fixtures/<id>.json`.
 3. Run `parseDiscogsRelease` on it (via `load-parser.ts`, which loads the userscript in a Node VM with the required mocks).
-4. Compare the output to `snapshots/<id>.json`.
+4. Compare the output to `snapshots/<id>.json` using `toMatchFileSnapshot`.
 
 CI runs `pnpm test:discogs` on every pull request. No Discogs API calls are made during normal test runs.
 
 ## Adding a test case
 
 1. Find a Discogs release that exercises the behaviour you care about.
-2. Add its API URL to `RELEASE_URLS` in `config.ts`:
+2. Add an entry to `RELEASES` in `config.ts`:
 
     ```ts
-    export const RELEASE_URLS = [
+    export const RELEASES = [
         // ...
-        'https://api.discogs.com/releases/1234567', // brief note on what this covers
+        { url: 'https://api.discogs.com/releases/1234567', description: 'brief note on what this covers' },
     ] as const;
     ```
 
     Use the API URL (`https://api.discogs.com/releases/<id>`), not the website URL.
 
-3. Fetch fixtures and generate snapshots (see below).
+3. Fetch the fixture and generate the snapshot (see below).
 4. Commit the new fixture, snapshot, and config change.
 
 Pick releases that cover distinct edge cases (multi-disc tracklists, unusual side numbering, nested sub-tracks, etc.) rather than many similar ones.
@@ -57,14 +60,18 @@ Run all tests:
 pnpm test:discogs
 ```
 
-Fetch fixtures from the Discogs API and regenerate all snapshots (needed after adding a release or after an intentional parser change):
+Fetch fixtures from the Discogs API and regenerate all snapshots (needed after adding a release):
 
 ```bash
-pnpm test:discogs:update
+pnpm test:discogs:update-fixtures
 ```
 
-`pnpm test` currently is an alias for `pnpm test:discogs` since there are no other tests.
+Approve updated snapshots after an intentional parser change (no network request):
 
-After adding a URL to `config.ts`, run `pnpm test:discogs:update` once to create its fixture and snapshot, then `pnpm test:discogs` to confirm everything passes.
+```bash
+./node_modules/.bin/vitest run tests/discogs -u
+```
 
-If a test fails after a parser change you intended, review the diff, then run `pnpm test:discogs:update` to refresh the snapshots and commit the updated files.
+After adding an entry to `config.ts`, run `pnpm test:discogs:update-fixtures` once to create its fixture and snapshot, then `pnpm test:discogs` to confirm everything passes.
+
+If a test fails after a parser change you intended, review the diff, then run `vitest -u` to approve the new snapshots and commit the updated files.
