+++ tests/discogs/config.ts
@@ -1,13 +1,12 @@
-/** Discogs API release URLs to test. */
-export const RELEASE_URLS = [
-    'https://api.discogs.com/releases/1996829', // tracks grouped under a header track
-    'https://api.discogs.com/releases/1156598', // unusual LP side track numbering
-    'https://api.discogs.com/releases/5880212', // multi disc tracksets
-    'https://api.discogs.com/releases/350350', // single header track
-    'https://api.discogs.com/releases/15313328', // complex multi-disc tracksets
-    'https://api.discogs.com/releases/3315779', // standard A/B sides LP
-    'https://api.discogs.com/releases/15771199', // CD+CD+DVD
-    'https://api.discogs.com/releases/1450895', // A/B+C/D with incorrectly nested subtracks
-    'https://api.discogs.com/releases/8661943', // A/B+C/D with correctly nested subtracks
-    'https://api.discogs.com/releases/16345092', // split track on second CD
+export const RELEASES = [
+    { url: 'https://api.discogs.com/releases/1996829', description: 'tracks grouped under a header track' },
+    { url: 'https://api.discogs.com/releases/1156598', description: 'unusual LP side track numbering' },
+    { url: 'https://api.discogs.com/releases/5880212', description: 'multi disc tracksets' },
+    { url: 'https://api.discogs.com/releases/350350', description: 'single header track' },
+    { url: 'https://api.discogs.com/releases/15313328', description: 'complex multi-disc tracksets' },
+    { url: 'https://api.discogs.com/releases/3315779', description: 'standard A/B sides LP' },
+    { url: 'https://api.discogs.com/releases/15771199', description: 'CD+CD+DVD' },
+    { url: 'https://api.discogs.com/releases/1450895', description: 'A/B+C/D with incorrectly nested subtracks' },
+    { url: 'https://api.discogs.com/releases/8661943', description: 'A/B+C/D with correctly nested subtracks' },
+    { url: 'https://api.discogs.com/releases/16345092', description: 'split track on second CD' },
 ] as const;
