+++ tests/discogs/snapshots/3315779.json
@@ -263,5 +263,7 @@
             "link_type": 76
         }
     ],
-    "secondary_types": ["compilation"]
+    "secondary_types": [
+        "compilation"
+    ]
 }
