+++ tests/discogs/snapshots/5880212.json
@@ -93,6 +93,8 @@
             "link_type": 76
         }
     ],
-    "secondary_types": ["compilation"],
+    "secondary_types": [
+        "compilation"
+    ],
     "barcode": "028946003121"
 }
